js code cleanup
