+++ html/login.html
@@ -65,8 +65,7 @@
             <button class="close-btn" id="popup-close-btn">확인</button>
         </div>
     </div>
-    <script src="../js/login.js"></script>
-    <script src="../common/login_signup.js"></script>
+    <script src="../js/login.js" type="module"></script>
 </body>
 
 </html>
