+++ html/signup.html
@@ -80,8 +80,7 @@
             <button class="close-btn" id="popup-close-btn">확인</button>
         </div>
     </div>
-    <script src="../js/signup.js"></script>
-    <script src="../common/login_signup.js"></script>
+    <script src="../js/signup.js" type="module"></script>
 </body>
 
 </html>
